HZN-605: Modified the method longValue to have Integer argument

+++ features/events/traps/src/main/java/org/opennms/netmgt/trapd/TrapdIpMgr.java
@@ -79,6 +79,11 @@ public interface TrapdIpMgr {
      */
     long removeNodeId(String addr);
 
-    long longValue(Long result);
+    /**
+     * Method will check value for null. 
+     * @param result
+     * @return If result is null returns -1 else returns result itself 
+     */
+    long longValue(Integer result);
 
 }
